Cleanup: Rename UserHub components
