+++ src/components/common/Extensions/UserHub/UserHub.tsx
@@ -8,7 +8,7 @@ import Select from '~v5/common/Fields/Select/index.ts';
 import TitleLabel from '~v5/shared/TitleLabel/index.ts';
 
 import { tabList } from './consts.ts';
-import ReputationTab from './partials/ReputationTab/index.ts';
+import BalanceTab from './partials/BalanceTab/index.ts';
 import StakesTab from './partials/StakesTab/index.ts';
 import TransactionsTab from './partials/TransactionsTab/index.ts';
 import { type UserHubProps, UserHubTabs } from './types.ts';
@@ -113,7 +113,7 @@ const UserHub: FC<UserHubProps> = ({
       </div>
       <div className={clsx('relative h-full w-full')}>
         {selectedTab === UserHubTabs.Balance && (
-          <ReputationTab onTabChange={handleTabChange} />
+          <BalanceTab onTabChange={handleTabChange} />
         )}
         {selectedTab === UserHubTabs.Stakes && <StakesTab />}
         {selectedTab === UserHubTabs.Transactions && (
