+++ src/components/common/Extensions/UserHub/partials/BalanceTab/BalanceTab.styles.ts
