+++ src/components/common/Extensions/UserHub/partials/BalanceTab/index.ts
@@ -1,3 +1,3 @@
-export { default } from './ReputationTab.tsx';
+export { default } from './BalanceTab.tsx';
 export { default as TotalReputation } from './partials/TotalReputation.tsx';
 export { default as PendingReputation } from './partials/PendingReputation/index.ts';
